feat: centralize checkout session logic
